Created conftest.py as per the comment.
